benchmark_fix WIP

+++ src/benchmarks/benchmark.h
@@ -197,7 +197,7 @@ struct benchmark_clo
  */
 struct worker_info
 {
-	unsigned index;			/* index of worker thread */
+	size_t index;			/* index of worker thread */
 	struct operation_info *opinfo;	/* operation info structure */
 	size_t nops;			/* number of operations */
 	void *priv;			/* worker's private data */
@@ -210,7 +210,7 @@ struct operation_info
 {
 	struct worker_info *worker;	/* worker's info */
 	struct benchmark_args *args;	/* benchmark arguments */
-	unsigned index;			/* operation's index */
+	size_t index;			/* operation's index */
 	benchmark_time_t t_diff;	/* timestamp of start */
 };
